requirements.txt
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-cloud-texttospeech==2.14.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gTTS==2.3.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.4
WTForms==3.0.1
